De-duplicate mouse scroll events
This bug is caused by a deprecated pointer event pointer::axis_discrete.

This event is received additionaly to pointer::axis, and we add them
together. Meaning one scroll event, with a STEP 15, plus a discrete
step, equal to two event that we will sent to the VNC server.

pointer::axis_discrete is deprecated as version 8, and is always coupled
with pointer::axis. Also, it is not sent by touchpad scrolls.

https://wayland.app/protocols/wayland#wl_pointer:event:axis_discrete

We can safely ignore it.