Rework the image composition using wl_subcompositor and wp_viewporter
This offload the scaling and the rendering to the compositor, and
it simplify our event handling code because we don't have to compute
internal coordinate anymore.

Because we now scale using the compositor viewporter, the buffers aren't
stretched multiple times in case of output scaling. We also don't have
to know the fractional scale value at all. The VNC client buffers are
dimensioned using the server values, and we just viewport them to
the surface.

Also, now the input inhibitor only trigger when the pointer enter
the subsurface, and not the background surface.

The most surprising change is arround the pointer handling. Now that
we have multiple wl_surface, we have to ignore events coming from the
background one. The pointers can enter and leave both surface, so we
just ignore events when we in the background one.