Fix held keys issues while focusing or unfocusing wlvncc
At the moment there is big issues when the user hold some keys while
entering or leaving the wlvncc top level. More dramatically when holding
modifiers keys. The internal xkb_state was still properly updated, so
it causes even more problems if CAPS_LOCK was locked in, and released
out.

This patch seems to solves all the issues I encountered.

To handle this properly, we send keyboard released events while
unfocusing, and pressed events while focusing.

To do so we have to keep track of the pressed keys. There is afaik no
way to use the xkb_state for this, because it only track the modifiers
and layouts states.