Make the shortcut inhibitor to works on the bg surface too
The fact it works on the internal surface, but not on the bg surface
is a really confusing behavior. It can really feels like it is bugged,
in case some of the internal buffer is also black (example looking
at a video with an aspect ratio not fitting in the display).

On the first place, I did this because I thought we had to because of
the F12 toggle key. But this toggle also is triggerable on the bg
surface so.