move inhibitor on keyboard.{enter,leave}
I think this make more sense. For users using keyboard based windows
manager, they can have full access to the remote with just keyboard. And
use F12 to detach from it. For users using regular desktop environment,
they have to actually click on the wlvncc top level to inhibit their
shortcuts.

This make the inhibitor to works on the whole toplevel, covering the
bg too. And this make the F12 toggle button more reliable, because
moving the pointer from and out of the intern surface, to the bg
one, to not restart the inhibitor.