Do not destroy and recreate shortcut inhibitors on focus/unfocus
This make F12 and the compositor --inhibited bindsyms more consistent,
because it toggle the inhibitors permanently.

We now inhibitor_inhibit immediately on the inhibitor_init, instead
of waiting for some enter events from pointer or keyboard. We also
inhibitor_inhibit from _add_seat, in case a new seat is created
later.