Cleanup outdated rendering related code
I'm not really sure if this change bring something noticeable for the
user. But at least, this is theorically more robust, and the code
is more simple.

Drop outdated is_attached, please_clean_up, is_frame_committed.

Dropped the frame callback completely. There now is no value to keep it.