Add script to connect to server over ssh tunnel
It is recommended the server only allowed connections from localhost.
Therefore a ssh port forwarding connection to the local port on the
server is required.

This script opens a port forwarding connection to the server via ssh.
wlvncc then connects to this local port.
When the script terminates the ssh connection is closed.

Closes: #32